% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfit.R
\docType{package}
\name{zfit}
\alias{zfit}
\title{zfit: Fit models in a pipe}
\description{
The goal of \code{zfit} is to improve the usage of basic model
fitting functions within a piped work flow, in particular when
passing and processing a tibble (or data.frame) using
\code{dplyr} and associated packages.
}
