% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfit_base_zfunction_zfitter.R
\name{zfitter}
\alias{zfitter}
\title{Create a pipe-friendly version of a given fitting function}
\usage{
zfitter(fun)
}
\arguments{
\item{fun}{The fitting function to adapt. The name should not be quoted,
rather, the actual function should be passed (prefixed with package
if needed)}
}
\description{
Compared to just using \code{zfunction()}, this function stores the base
name, allowing one to use full name of the original fitting function
(such as \code{MASS::polr}), which is useful to just pull a single
fitting function from a package without loading it. It also shortens
very long data names (longer than 32 characters by default), which
otherwise are a nuisance when the data comes from the pipe, because
the pipeline gets converted to a very long function call.
}
