% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfit.R
\name{zglm}
\alias{zglm}
\title{Run a glm model in a pipe (see \code{zlm})}
\usage{
zglm(data, formula, family = gaussian, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the model data.}

\item{formula}{The \code{formula} to be fitted.}

\item{family}{The \code{family} function to use for fitting the model.}

\item{...}{Other arguments to be passed to the \code{glm} function.}
}
\value{
A fitted model.
}
\description{
Run a glm model in a pipe (see \code{zlm})
}
\seealso{
Other zfit: 
\code{\link{zlm}()},
\code{\link{zlogit}()},
\code{\link{zprint}()},
\code{\link{zprobit}()}
}
\concept{zfit}
