% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfit.R
\name{zpoisson}
\alias{zpoisson}
\title{Run a poisson model in a pipe (see \code{zlm})}
\usage{
zpoisson(data, formula, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the model data.}

\item{formula}{The \code{formula} to be fitted.}

\item{...}{Other arguments to be passed to the \code{glm} function.}
}
\value{
A fitted model.
}
\description{
Run a poisson model in a pipe (see \code{zlm})
}
\seealso{
Other zfit: 
\code{\link{zglm}()},
\code{\link{zlm_robust}()},
\code{\link{zlm}()},
\code{\link{zlogit}()},
\code{\link{zprint}()},
\code{\link{zprobit}()}
}
\concept{zfit}
